Implemented first animation example for LineSeries
