+++ Source/Examples/WPF/WpfExamples/Examples/AnimationsDemo/ViewModels/Interfaces/IAnimationViewModel.cs
@@ -1,0 +1,17 @@
+﻿// --------------------------------------------------------------------------------------------------------------------
+// <copyright file="IAnimationViewModel.cs" company="OxyPlot">
+//   Copyright (c) 2014 OxyPlot contributors
+// </copyright>
+// --------------------------------------------------------------------------------------------------------------------
+
+
+namespace AnimationsDemo
+{
+    using System;
+
+    public interface IAnimationViewModel
+    {
+        void Animate();
+        void Animate(IEasingFunction easingFunction, TimeSpan duration);
+    }
+}
