+++ Source/Examples/WPF/WpfExamples/Examples/AnimationsDemo/Views/LineSeriesView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="AnimationsDemo.LinearBarView"
+﻿<UserControl x:Class="AnimationsDemo.LineSeriesView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
