+++ Source/Examples/WPF/WpfExamples/Examples/AnimationsDemo/Views/LineSeriesView.xaml.cs
@@ -1,0 +1,17 @@
+﻿// --------------------------------------------------------------------------------------------------------------------
+// <copyright file="LineSeriesView.xaml.cs" company="OxyPlot">
+//   Copyright (c) 2014 OxyPlot contributors
+// </copyright>
+// --------------------------------------------------------------------------------------------------------------------
+
+namespace AnimationsDemo
+{
+    public partial class LineSeriesView
+    {
+        public LineSeriesView()
+        {
+            this.InitializeComponent();
+            this.DataContext = new LineSeriesViewModel();
+        }
+    }
+}
